Change Directory structure
